[FIX] remove filter[has_delivery_order] SO & update role

+++ app/Http/Controllers/Api/RoleController.php
@@ -37,6 +37,7 @@ class RoleController extends Controller
      */
     public function store(RoleStoreRequest $request)
     {
+        abort_if(!auth()->user()->tokenCan('role_create'), 403);
         $role = new Role();
         $role->name = $request->name;
         $role->guard_name = 'web';
@@ -66,6 +67,8 @@ class RoleController extends Controller
      */
     public function update(Role $role, RoleUpdateRequest $request)
     {
+        abort_if(!auth()->user()->tokenCan('role_edit'), 403);
+        if ($role->id == 1) return response()->json(['message' => 'Role admin can not updated!']);
         $role->name = $request->input('name');
         $role->save();
         $role->syncPermissions($request->permission_ids ?? []);
@@ -81,6 +84,7 @@ class RoleController extends Controller
     public function destroy(Role $role)
     {
         abort_if(!auth()->user()->tokenCan('role_delete'), 403);
+        if ($role->id == 1) return response()->json(['message' => 'Role admin can not deleted!']);
         $role->delete();
         return $this->deletedResponse();
     }
